Fixed validation; its working properly now
The reason was - we did everything fine with adding all the
properties to Form components, like ref, onChange, formValue etc.
but after doing that we could not type anything to fields.
Thats because when we were setting the state we changed wrong key in the
state object - formValue was changed, and our key is named formData .___.